small cleanup
remove the unlink_socket variable and it's check as they wher
always true, as error and the end of the programm can only be
reached when the socket is already set up.
Also unlinking an non existent file does not result in any error.

also some whitespace cleanup in the surrounding area.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>