replace service_config_exists with has_services
Remove the "service_config_exists" method from the Environment class
and replace it with a new method in PVE::HA::Tools.
First we do not need this method in the Environment class and second
this can be reused more for checking if a note has any service
configured.

This moves the regression test also a little closer to the real
world behaviour (while maintaining determinism).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>