improve verbosity of API status call
Display the state of a LRM when it is not in the "active" state
and has some service configured.

This should reduce confiusion when the LRM is active but still has
to wait for its lock. In such a case the user only sees that it's
active and could think it's malfunctioning because no action
happens.

Also displays the LRM mode, so that we can see when we restart the LRM.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Signed-off-by: Dietmar Maurer <dietmar@proxmox.com>