use correct verify function for ha-group-node-list
We use pve_verify_nodename but we expect here a value with the
syntax: nodename[:priority]
So if someone used the optional priority this failed unecessarily.

select_service_node in the Manager class expects this value and
splits it itself into nodename and priority so we can just return
it if it matches the ha-group-node-list format.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>